CHG: Moved quaternion based tests to quaternion folder
